pyproject.toml
[tool.poetry]
name = "stockbot"
version = "0.1.0"
description = ""
authors = ["namig <guseinovnamig41@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-telegram-bot = "^21.8"
django-environ = "^0.11.2"
pydantic-settings = "^2.6.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
